// src/constants/enums.ts
export enum PlaylistTag {
  Carousel = 'carousel',
  TopPicks = 'top-picks',
}

export enum UserRole {
  Admin = 'admin',
  User = 'user',
}

export enum Utilization {
  WithPatients = 'With Patients/Clients',
  WithStudents = 'With Students',
  Personal = 'Personal Use',
  Other = 'Other',
}

export enum Interests {
  ADDICTION_RECOVERY = 'Addiction Recovery',
  PTSD_GENERAL_TRAUMA = 'PTSD / General Trauma',
  GRIEF_DEATH = 'Grief / Death',
  OVERDOSE_SUICIDE = 'Overdose / Suicide',
  DEPRESSION_ANXIETY = 'Depression / Anxiety',
  SEX_RELATIONSHIPS = 'Sex & Relationships',
  OTHER = 'Other',
}